Adding Exception Handling

+++ MusicManagementSystem/Controllers/SongsController.cs
@@ -22,42 +22,72 @@ namespace MusicManagementSystem.Controllers
         [HttpGet]
         public IActionResult Get()
         {
-            return _dbContext.Songs;
+            // return _dbContext.Songs;
+
+            return StatusCode(StatusCodes.Status200OK);
         }
 
         // GET api/<SongsController>/5
         [HttpGet("{id}")]
-        public Song Get(int id)
+        public IActionResult Get(int id)
         {
             var song =  _dbContext.Songs.Find(id);
-            return song;
+            if (song == null)
+            {
+                return NotFound("No record found in this Id");
+            }
+            else
+            {
+                return Ok(song);
+            }
+                
         }
 
         // POST api/<SongsController>
         [HttpPost]
-        public void Post([FromBody] Song song)
+        public IActionResult Post([FromBody] Song song)
         {
             _dbContext.Songs.Add(song);
             _dbContext.SaveChanges();
+            return StatusCode(StatusCodes.Status201Created);
         }
 
         // PUT api/<SongsController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Song songObj)
+        public IActionResult Put(int id, [FromBody] Song songObj)
         {
            var song = _dbContext.Songs.Find(id);
-            song.Title = songObj.Title;
-            song.Language = songObj.Language;
-            _dbContext.SaveChanges();
+
+            if(song == null)
+            {
+                return NotFound("No record found in this Id");
+            }else
+            {
+                song.Title = songObj.Title;
+                song.Language = songObj.Language;
+                _dbContext.SaveChanges();
+                return Ok("Record Updated Successfully");
+            }
+            
+         
         }
 
         // DELETE api/<SongsController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
            var song = _dbContext.Songs.Find(id);
-           _dbContext.Songs.Remove(song);
-           _dbContext.SaveChanges();
+            if (song == null)
+            {
+                return NotFound("No record found in this Id");
+            }
+            else
+            {
+                _dbContext.Songs.Remove(song);
+                _dbContext.SaveChanges();
+                return Ok("Record Deleted");
+            }
+                
         }
     }
 }
